Extract shared OAuth callback server machinery
